# openapi.yaml
openapi: 3.0.3
info:
  title: YOLO Training Visualization Platform API
  version: 1.0.0
  description: |
    REST API for YOLO Training Visualization Platform - A comprehensive tool for dataset management, 
    model training, testing, validation, export, and TCP-based image processing.
    
    ## Modules
    - **IDataset**: Dataset management and Label Studio integration
    - **ITraining**: Training task creation, execution and monitoring
    - **IModel**: Model testing, validation, export and Triton deployment
    - **IImageProcessor**: TCP image processing service integration
    
  contact:
    name: chzane
    email: slxzane@outlook.com
    url: https://github.com/chzane/YoloTrainingVisualizationPlatform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:10799
    description: Local development server
paths:
...
  /IDataset/uploadDataset:
    post:
      summary: Upload a dataset archive and register it
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                description:
                  type: string
                version:
                  type: string
                  example: v1.0.0
                type:
                  type: string
                  description: Dataset type
                  enum: [YOLO, COCO]
                include_yaml:
                  type: string
                  description: '"1" to auto-detect YAML in archive; otherwise provide fields below'
                train:
                  type: string
                val:
                  type: string
                test:
                  type: string
                nc:
                  type: integer
                names:
                  type: string
      responses:
        '200':
          description: Upload status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /IDataset/deleteDataset:
    post:
      summary: Delete a dataset directory by absolute path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
              required: [path]
      responses:
        '200':
          description: Delete status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
...
  /IDataset/listLabelStudioProjects:
    get:
      summary: List Label Studio projects
      parameters:
        - in: query
          name: base_url
          schema: { type: string }
          required: true
          description: Label Studio server URL
        - in: query
          name: token
          schema: { type: string }
          required: false
          description: Label Studio API token (optional)
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
  /IDataset/buildDatasetFromLabelStudio:
    post:
      summary: Build YOLO dataset from Label Studio project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                base_url: { type: string, description: Label Studio server URL }
                token: { type: string, description: API token (optional) }
                project_id: { type: integer, description: Project ID }
                name: { type: string, description: Dataset name }
                version: { type: string, description: Dataset version }
                description: { type: string }
                exportType: { type: string, default: CSV }
                splits: { type: array, items: { type: number }, example: [0.8, 0.2, 0.0] }
                seed: { type: integer, default: 42 }
                download_images: { type: boolean, default: true }
                class_names: { type: array, items: { type: string }, nullable: true }
              required: [base_url, project_id, name, version]
      responses:
        '200':
          description: Dataset build status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StdResponse'
...
components:
  schemas:
    StdResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: OK
        data:
          description: Response payload (varies by endpoint)
          nullable: true